unit: stable serialization order
66542898b0f082300f28123eccc6bea01b4ec343 introduced consolidation of
UnitOptions into Sections during serialization. This preserved ordering
of individual options within sections, but since a map was used to house
the sections their order was nondeterministic.

This change moves to preserving the first-seen order of the sections, as
well as continuing to preserve the order of units with sections. For
example:

	[UnitOption{A, B, b}, UnitOption{C, D, d}, UnitOption{A, Z, z}]

will consistently be written as:

	[A]
	B=b
	Z=z
	[C]
	D=d