journal: split into read and write packages
0020698a980adf8a2c440d9be709edc85b54544a (via #110) introduced bindings
to read from the journal, which wrap around the sd-journal C API (the
recommended approach to access the journal [1]). Unfortunately that
means that the github.com/coreos/go-systemd/journal package suddenly
moved from being pure Go to having a hard dependency on cgo and
sd-journal, which caused issues on some platforms.

After thinking about this a bit more I think it makes most sense to
split apart the two (read and write) APIs into different packages (which
Zbigniew actually originally suggested a long time ago [3]), as the
implementations are completely separate. This also means we don't need
to futz about with build tag magic to allow people to e.g. only use the
pure-Go part of the API.

Fixes #115 and #116.

[1]: http://lists.freedesktop.org/archives/systemd-devel/2014-June/019904.html
[2]: https://github.com/coreos/go-systemd/issues/116
[3]: http://lists.freedesktop.org/archives/systemd-devel/2014-June/020067.html