dbus: make New() fall back to direct systemd connection
Previously, New() would attempt to connect to the
system-wide DBus. This changes the behaviour so
that in the event that such a connection fails, it falls
back to attempting a direct connection to the private
systemd socket (i.e. via NewSystemdConnection()).

A new, more specific `NewSystemConnection()` constructor
is introduced which provides the previous behaviour.

Fixes: #207