Fix TestCursor when cursor is valid
This fixes TestCursor so it returns an error when the ``cursor``
parameter is not the same as the current position.

From SD_JOURNAL_GET_CURSOR(3):
```
sd_journal_test_cursor() returns positive if the current entry
matches the specified cursor, 0 if it does not match the specified
cursor or a negative errno-style error code on failure.
    ```