Clarify documentation's notion of "current"
As per e.g. https://www.freedesktop.org/software/systemd/man/sd_journal_get_realtime_usec.html:

> Note that these functions will not work before sd_journal_next(3) (or related call) has been called at least once, in order to position the read pointer at a valid entry.

The prior wording of the documentation was ambiguous regarding its usage of "current". This change attempts to clarify that it's "current" in terms of "the state of the read pointer, adjusted by the Next/Previous family of functions" vs. "current" in terms of chronology.