activation: add two functions to provide listeners with names
Since v227, systemd can name each listener. This is useful when you
have different socket units activating the same service: you can't
really know in which order the socket will be provided. Systemd adds
an environment variable named LISTEN_FDNAMES for this purpose.

We encapsulate the name in the File object created from the file
descriptor. Unfortunately, once this is turned into a net.Listener,
there doesn't seem to have a way to fetch this information
back. Therefore, two new functions are provided to be able to easily
get named sockets.

The same could be done to PacketConns in the future.