activation: remove unsafe support for fd re-use
Package `activation` was previously offering an option to leave
environment variables untouched, in order to re-use fd at a future
point.
However, re-using fd is unsafe and results in racing against the GC
due to the way how `os.File` is implemented. Moreover, the Go runtime
does not (nor plan to) provide a safe interface for proper temporary
borrowing of existing fd.

Thus, this removes the `unsetEnv` option from all listeners (forcing
to always unset), and only keeps it where ownership can be transferred
back to the caller.