journal: SendVals with arbitrary auto-converting functions
This patch adds new function SendVals and Value interface which allows
the API users to use alternative syntax to provide journald values:

journal.SendVals(message, journal.PriInfo,
	journal.String("KEY", "VALUE"),
	journal.Int("INT_KEY", 1),
)

The whole suite of helper fucntions is provided, but they immediately
convert the value to string since carrying the type does not make much
sense in this case as the value is immediately written to byte buffer
and sent to journald. It is just a syntactic sugar I thought might be
useful, for my usecase only the journal.String is enough.