TestListJobs: fix flakiness
In some cases, TestListJobs fails like this:

	methods_test.go:712: oneshot.service job not found in list

This happens because 'ExecStart=/bin/sh sleep 400' doesn't really run
sleep; it merely fails. So, the job is queued and then ExecStart fails
instantly, and if the test checks the job status after the failure, the
test fails.

Fix the ExecStart.

While at it,
 - fix unit description;
 - remove ExecStartPost as it's useless (and also wrong).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>